pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "colorlog>=6.10.1",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.128.0",
    "gunicorn>=24.1.1",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pytest>=9.0.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "sqlmodel>=0.0.32",
]

[dependency-groups]
dev = ["mypy>=1.19.1", "pytest>=9.0.2", "ruff>=0.14.14"]

[tool.mypy]
python_version = "3.12"
files = ["app"]
strict = true
ignore_missing_imports = true
show_error_codes = true
pretty = true
cache_dir = ".mypy_cache"
plugins = ['pydantic.mypy']

[tool.mypy-pymongo]
ignore_missing_imports = true
follow_imports = false

[tool.ruff]
target-version = "py312"
src = ["app"]
fix = true
show-fixes = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]


[tool.pytest.ini_options]
minversion = "7.0"
pythonpath = ["."]
testpaths = ["./tests"]
python_files = ["test_*.py", "*_test.py"]
log_cli = true
log_level = "INFO"
addopts = [
    "-ra",
    "-v",
    "--tb=auto",
    "--color=auto",
    "--exitfirst",
    "--cov=app",
    "--cov-report=term-missing",
]
markers = [
    "unit: fast isolated tests",
    "integration: tests with external systems",
    "api: http contract tests",
]


[tool.alembic]
script_location = "%(here)s/alembic"

prepend_sys_path = ["."]

# timezone to use when rendering the date within the migration file
# as well as the filename.
# If specified, requires the tzdata library which can be installed by adding
# `alembic[tz]` to the pip requirements.
# string value is passed to ZoneInfo()
# leave blank for localtime
# timezone =

# max length of characters to apply to the "slug" field
# truncate_slug_length = 40

# set to 'true' to run the environment during
# the 'revision' command, regardless of autogenerate
# revision_environment = false

# set to 'true' to allow .pyc and .pyo files without
# a source .py file to be detected as revisions in the
# versions/ directory
# sourceless = false

# version location specification; This defaults
# to <script_location>/versions.  When using multiple version
# directories, initial revisions must be specified with --version-path.
# version_locations = [
#    "%(here)s/alembic/versions",
#    "%(here)s/foo/bar"
# ]


# set to 'true' to search source files recursively
# in each "version_locations" directory
# new in Alembic version 1.10
# recursive_version_locations = false

# the output encoding used when revision files
# are written from script.py.mako
output_encoding = "utf-8"

# This section defines scripts or Python functions that are run
# on newly generated revision scripts.  See the documentation for further
# detail and examples
# [[tool.alembic.post_write_hooks]]
# format using "black" - use the console_scripts runner,
# against the "black" entrypoint
# name = "black"
# type = "console_scripts"
# entrypoint = "black"
# options = "-l 79 REVISION_SCRIPT_FILENAME"
#
[[tool.alembic.post_write_hooks]]
# lint with attempts to fix using "ruff" - use the module runner, against the "ruff" module
name = "ruff"
type = "module"
module = "ruff"
options = "check --fix REVISION_SCRIPT_FILENAME"

[[tool.alembic.post_write_hooks]]
# Alternatively, use the exec runner to execute a binary found on your PATH
name = "ruff"
type = "exec"
executable = "ruff"
options = "check --fix REVISION_SCRIPT_FILENAME"
